+++ src/Requests/MoneyTransfer/CreateDisbursement.php
@@ -8,6 +8,7 @@ use Saloon\Contracts\Body\HasBody;
 use Saloon\Enums\Method;
 use Saloon\Http\Request;
 use Saloon\Traits\Body\HasJsonBody;
+use ZikraAuliya\FlipConnector\DTO\DisbursementPayload;
 
 /**
  * Create Disbursement
@@ -23,5 +24,24 @@ class CreateDisbursement extends Request implements HasBody
         return '/v3/disbursement';
     }
 
-    public function __construct() {}
+    public function __construct(
+        protected ?DisbursementPayload $payload = null
+    ) {}
+
+    /**
+     * Default Body
+     *
+     * @return array<int, int|string>
+     */
+    public function defaultBody(): array
+    {
+        if (! is_null($this->payload)) {
+            /** @var array<int, int|string> $payload */
+            $payload = $this->payload->toArray();
+
+            return $payload;
+        }
+
+        return [];
+    }
 }